MultiFlowSolver

Documentation for MultiFlowSolver.

MultiFlowSolver.add_lambda_to_rmlp

MultiFlowSolver.generate_planar_grid_mnfp

MultiFlowSolver.get_cross_count

MultiFlowSolver.get_used_arcs_and_cost

MultiFlowSolver.print_mnfp_graph

MultiFlowSolver.print_mnfp_instance

MultiFlowSolver.add_lambda_to_rmlp — Method

adds the lambda to the rmlp and ref array

MultiFlowSolver.generate_planar_grid_mnfp — Method

generate_planar_grid_mnfp()

Generate a planar MNFP instance based on a grid graph. Returns a named Tuple with output data.

MultiFlowSolver.get_cross_count — Method

returns how many times each long edge crosses another long edge

MultiFlowSolver.get_used_arcs_and_cost — Method

fills usedarcs with arcs used by newlambda_vertices and returns the cost

MultiFlowSolver.print_mnfp_graph — Method

saves a graph plot of the MNFP instance to a file

MultiFlowSolver.print_mnfp_instance — Method

print_mnfp_instance(mnfp::MnfpData)

Print a summary of the MNFP instance.